Refactor: split large UI files into focused components, hooks & lib modules
Pure structural refactor — no behavior change (typecheck + build clean).

lib:
- Split lib/api.ts → lib/ai.ts (AI client + SSE), lib/http.ts (errMessage);
  api.ts is now mock-CRUD only
- Add lib/err.ts (errText, was duplicated), lib/platform.ts (MOD),
  lib/live.ts (strip/statusClass/pathMatches/toggleIn)

hooks:
- useAssistantChat (chat engine out of AssistantPanel)
- useLatestRelease (out of SettingsModal)
- useSearchSuggest (ghost path-completion out of App)
- useLiveLogs (polling out of LiveView)

components:
- SettingsModal → settings/{ThemeTab,ImportTab,ApiKeysTab} (519 → 157 lines)
- App → catalog/{SearchBar,CatalogList,ShortcutBar} (365 → 233)
- LiveView → live/{StatusBadge,PathFilter,LiveRow} (532 → 285)
- AssistantPanel now presentation-only (256 → 162)
