+++ mocktail-ui/src/components/SettingsModal.tsx
@@ -1,71 +1,16 @@
 import { useEffect, useState } from 'react'
-import {
-  deleteAIKey,
-  fetchAIConfig,
-  fetchAIModels,
-  fetchAIProviders,
-  importMocks,
-  saveAIConfig,
-  type AIConfig,
-  type AIModel,
-  type AIProvider,
-  type ImportResult,
-} from '../lib/api'
-import { ACCENTS, type Accent, type Theme } from '../lib/theme'
+import { type Accent, type Theme } from '../lib/theme'
+import { useLatestRelease } from '../hooks/useLatestRelease'
 import { FaqAccordion } from './FaqAccordion'
+import { ThemeTab } from './settings/ThemeTab'
+import { ImportTab } from './settings/ImportTab'
+import { ApiKeysTab } from './settings/ApiKeysTab'
 
 export type SettingsTab = 'theme' | 'import' | 'apikeys' | 'help'
 
-const REPO = 'Huseyinnurbaki/mocktail'
 // Landing page. Update-available sends users here for install/upgrade docs, not the raw GitHub
 // release. The `#install` section is the contract — keep it on the landing page.
 const SITE_URL = 'https://getmocktail.com/#install'
-
-type ReleaseState = { status: 'checking' | 'latest' | 'outdated' | 'unknown'; latest?: string; url?: string }
-// Session cache — check GitHub at most once per app run (and never block anything on it).
-let releaseCache: ReleaseState | null = null
-
-function cmpVer(a: string, b: string): number {
-  const pa = a.replace(/^v/, '').split('.').map((n) => parseInt(n, 10) || 0)
-  const pb = b.replace(/^v/, '').split('.').map((n) => parseInt(n, 10) || 0)
-  for (let i = 0; i < Math.max(pa.length, pb.length); i++) {
-    const d = (pa[i] ?? 0) - (pb[i] ?? 0)
-    if (d !== 0) return d
-  }
-  return 0
-}
-
-/** Checks GitHub's latest release once (on first Settings open) and compares to the running version. */
-function useLatestRelease(): ReleaseState {
-  const [state, setState] = useState<ReleaseState>(releaseCache ?? { status: 'checking' })
-  useEffect(() => {
-    if (releaseCache) return
-    let alive = true
-    fetch(`https://api.github.com/repos/${REPO}/releases/latest`, {
-      headers: { Accept: 'application/vnd.github+json' },
-    })
-      .then((r) => (r.ok ? r.json() : Promise.reject(new Error(String(r.status)))))
-      .then((d: { tag_name?: string; html_url?: string }) => {
-        const tag = (d.tag_name ?? '').trim()
-        releaseCache = tag
-          ? {
-              status: cmpVer(tag, __APP_VERSION__) > 0 ? 'outdated' : 'latest',
-              latest: tag.replace(/^v/, ''),
-              url: d.html_url,
-            }
-          : { status: 'unknown' }
-        if (alive) setState(releaseCache)
-      })
-      .catch(() => {
-        releaseCache = { status: 'unknown' }
-        if (alive) setState(releaseCache)
-      })
-    return () => {
-      alive = false
-    }
-  }, [])
-  return state
-}
 
 const TABS: { id: SettingsTab; label: string }[] = [
   { id: 'theme', label: 'Theme' },
@@ -207,313 +152,6 @@ export function SettingsModal({
           </div>
         </div>
       </div>
-    </div>
-  )
-}
-
-function ThemeTab({
-  theme,
-  setTheme,
-  accent,
-  setAccent,
-}: {
-  theme: Theme
-  setTheme: (t: Theme) => void
-  accent: Accent
-  setAccent: (a: Accent) => void
-}) {
-  const opts: { id: Theme; label: string; desc: string }[] = [
-    { id: 'system', label: 'System', desc: 'Match your OS appearance' },
-    { id: 'light', label: 'Light', desc: 'Always light' },
-    { id: 'dark', label: 'Dark', desc: 'Always dark' },
-  ]
-  return (
-    <div className="flex flex-col gap-4">
-      <div className="flex flex-col gap-2">
-        <div className="text-[11px] font-semibold uppercase tracking-[0.06em] text-muted">Mode</div>
-        {opts.map((o) => (
-          <button
-            key={o.id}
-            onClick={() => setTheme(o.id)}
-            className={`flex items-center justify-between rounded-[9px] border px-3 py-2 text-left ${
-              theme === o.id ? 'border-accent bg-accent-tint' : 'border-border hover:bg-surface-sunken'
-            }`}
-          >
-            <span>
-              <span className="text-[13px] font-medium">{o.label}</span>
-              <span className="block text-[12px] text-muted">{o.desc}</span>
-            </span>
-            {theme === o.id && <span className="text-accent-text">✓</span>}
-          </button>
-        ))}
-      </div>
-
-      <div className="flex flex-col gap-2">
-        <div className="flex items-baseline justify-between">
-          <div className="text-[11px] font-semibold uppercase tracking-[0.06em] text-muted">Accent</div>
-          <div className="text-[12px] text-muted">
-            {ACCENTS.find((a) => a.id === accent)?.label}
-          </div>
-        </div>
-        <div className="flex flex-wrap gap-3 rounded-[11px] border border-border bg-surface-sunken px-3.5 py-3">
-          {ACCENTS.map((a) => (
-            <button
-              key={a.id}
-              onClick={() => setAccent(a.id)}
-              title={a.label}
-              aria-label={a.label}
-              aria-pressed={accent === a.id}
-              className={`flex h-[32px] w-[32px] items-center justify-center rounded-full ring-offset-2 ring-offset-surface-sunken transition ${
-                accent === a.id
-                  ? 'ring-2 ring-accent'
-                  : 'ring-1 ring-transparent hover:scale-110 hover:ring-fg/25'
-              }`}
-            >
-              <span
-                className="flex h-[22px] w-[22px] items-center justify-center rounded-full text-white shadow-sm ring-1 ring-inset ring-black/10"
-                style={{ background: `oklch(0.68 0.15 ${a.hue})` }}
-              >
-                {accent === a.id && (
-                  <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="3.5" strokeLinecap="round" strokeLinejoin="round">
-                    <polyline points="20 6 9 17 4 12" />
-                  </svg>
-                )}
-              </span>
-            </button>
-          ))}
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function ImportTab({ onImported }: { onImported: () => void }) {
-  const [text, setText] = useState('')
-  const [busy, setBusy] = useState(false)
-  const [result, setResult] = useState<ImportResult | null>(null)
-  const [err, setErr] = useState<string | null>(null)
-
-  async function run() {
-    setBusy(true)
-    setErr(null)
-    setResult(null)
-    try {
-      const r = await importMocks(text)
-      setResult(r)
-      onImported()
-    } catch (e: unknown) {
-      setErr(e instanceof Error ? e.message : String(e))
-    } finally {
-      setBusy(false)
-    }
-  }
-
-  function onFile(e: React.ChangeEvent<HTMLInputElement>) {
-    const f = e.target.files?.[0]
-    if (f) void f.text().then(setText)
-  }
-
-  return (
-    <div className="flex flex-col gap-3 text-[13px]">
-      <div className="text-[12.5px] text-muted">
-        Paste exported mocks JSON (an array or <span className="font-mono">{'{ "Apis": [...] }'}</span>),
-        or choose a file. Existing paths are skipped.
-      </div>
-      <input type="file" accept="application/json,.json" onChange={onFile} className="text-[12px]" />
-      <textarea
-        value={text}
-        onChange={(e) => setText(e.target.value)}
-        rows={8}
-        placeholder={'[ { "Method": "GET", "Endpoint": "/api/v1/users", "Response": { } } ]'}
-        className="w-full resize-none rounded-[8px] border border-border bg-surface-sunken p-3 font-mono text-[12px] outline-none"
-      />
-      <button
-        onClick={() => void run()}
-        disabled={busy || !text.trim()}
-        className="h-[32px] rounded-[8px] bg-accent text-[13px] font-semibold text-accent-on disabled:opacity-40"
-      >
-        {busy ? 'Importing…' : 'Import mocks'}
-      </button>
-      {err && <div className="text-[12.5px] text-error">{err}</div>}
-      {result && (
-        <div className="text-[12.5px] text-muted">
-          Imported <span className="text-success">{result.imported}</span> · skipped {result.skipped}{' '}
-          · failed {result.failed}
-        </div>
-      )}
-    </div>
-  )
-}
-
-function errText(e: unknown): string {
-  return e instanceof Error ? e.message : String(e)
-}
-
-function ApiKeysTab() {
-  const [config, setConfig] = useState<AIConfig | null>(null)
-  const [providers, setProviders] = useState<AIProvider[]>([])
-  const [models, setModels] = useState<AIModel[]>([])
-  const [keyInput, setKeyInput] = useState('')
-  const [busy, setBusy] = useState(false)
-  const [err, setErr] = useState<string | null>(null)
-  const [note, setNote] = useState<string | null>(null)
-
-  async function refresh(withModels: boolean) {
-    const [cfg, prov] = await Promise.all([fetchAIConfig(), fetchAIProviders()])
-    setConfig(cfg)
-    setProviders(prov.providers)
-    if (withModels && cfg.configured) {
-      try {
-        setModels((await fetchAIModels()).models)
-      } catch {
-        /* leave prior models; dropdown just won't refresh */
-      }
-    }
-  }
-
-  useEffect(() => {
-    refresh(true).catch((e) => setErr(errText(e)))
-  }, [])
-
-  const envManaged = config?.source === 'env'
-  const configured = config?.configured ?? false
-
-  async function saveKey() {
-    const key = keyInput.trim()
-    if (!key || busy) return
-    setBusy(true)
-    setErr(null)
-    setNote(null)
-    try {
-      await saveAIConfig({ apiKey: key })
-      setKeyInput('')
-      setNote('Key validated and saved.')
-      await refresh(true)
-    } catch (e) {
-      setErr(errText(e))
-    } finally {
-      setBusy(false)
-    }
-  }
-
-  async function removeKey() {
-    setBusy(true)
-    setErr(null)
-    setNote(null)
-    try {
-      await deleteAIKey()
-      setModels([])
-      await refresh(false)
-    } catch (e) {
-      setErr(errText(e))
-    } finally {
-      setBusy(false)
-    }
-  }
-
-  async function pick(input: { provider?: string; model?: string }) {
-    setBusy(true)
-    setErr(null)
-    setNote(null)
-    try {
-      await saveAIConfig(input)
-      await refresh(!!input.provider) // a provider change re-fetches models
-    } catch (e) {
-      setErr(errText(e))
-    } finally {
-      setBusy(false)
-    }
-  }
-
-  return (
-    <div className="flex flex-col gap-3 text-[13px]">
-      <div className="text-[12.5px] leading-[1.55] text-muted">
-        Your key is stored on this machine (OS keychain) and used only server-side — it never leaves
-        your backend. Powers the ✨ Assistant chat.
-      </div>
-
-      {/* Provider */}
-      <div className="flex flex-col gap-1">
-        <label className="text-[12.5px] text-muted">Provider</label>
-        <select
-          value={config?.provider ?? ''}
-          disabled={busy || providers.length <= 1}
-          onChange={(e) => pick({ provider: e.target.value })}
-          className="h-[34px] rounded-[8px] border border-border bg-surface px-2 text-[13px] disabled:opacity-60"
-        >
-          {providers.map((p) => (
-            <option key={p.id} value={p.id}>
-              {p.name}
-            </option>
-          ))}
-        </select>
-      </div>
-
-      {/* API key */}
-      <div className="flex flex-col gap-1">
-        <label className="text-[12.5px] text-muted">API key</label>
-        {envManaged ? (
-          <div className="rounded-[8px] border border-border bg-surface-sunken px-3 py-2 text-[12.5px] text-muted">
-            Managed via <span className="font-mono">MOCKTAIL_AI_API_KEY</span> — read-only here.
-            {config?.keyHint && <span className="ml-1 font-mono">({config.keyHint})</span>}
-          </div>
-        ) : configured ? (
-          <div className="flex items-center justify-between gap-2 rounded-[8px] border border-border bg-surface-sunken px-3 py-2">
-            <span className="text-[12.5px] text-muted">
-              Key set <span className="ml-1 font-mono text-fg">{config?.keyHint}</span>
-            </span>
-            <button
-              onClick={removeKey}
-              disabled={busy}
-              className="text-[12px] text-error hover:underline disabled:opacity-40"
-            >
-              Remove
-            </button>
-          </div>
-        ) : (
-          <div className="flex gap-2">
-            <input
-              value={keyInput}
-              onChange={(e) => setKeyInput(e.target.value)}
-              onKeyDown={(e) => e.key === 'Enter' && saveKey()}
-              type="password"
-              placeholder="sk-ant-…"
-              className="h-[34px] flex-1 rounded-[8px] border border-border bg-surface px-2 font-mono text-[13px]"
-            />
-            <button
-              onClick={saveKey}
-              disabled={busy || !keyInput.trim()}
-              className="h-[34px] shrink-0 rounded-[8px] bg-accent px-3 text-[13px] font-semibold text-accent-on disabled:opacity-40"
-            >
-              {busy ? 'Saving…' : 'Save'}
-            </button>
-          </div>
-        )}
-      </div>
-
-      {/* Model — only meaningful once a key is set */}
-      {configured && (
-        <div className="flex flex-col gap-1">
-          <label className="text-[12.5px] text-muted">Model</label>
-          <select
-            value={config?.model ?? ''}
-            disabled={busy || models.length === 0}
-            onChange={(e) => pick({ model: e.target.value })}
-            className="h-[34px] rounded-[8px] border border-border bg-surface px-2 text-[13px] disabled:opacity-60"
-          >
-            {models.length === 0 && <option value={config?.model ?? ''}>{config?.model}</option>}
-            {models.map((m) => (
-              <option key={m.id} value={m.id}>
-                {m.displayName}
-                {m.recommended ? ' · recommended' : ''}
-              </option>
-            ))}
-          </select>
-        </div>
-      )}
-
-      {err && <div className="text-[12.5px] text-error">{err}</div>}
-      {note && <div className="text-[12.5px] text-success">{note}</div>}
     </div>
   )
 }
